@@ -24,13 +24,13 @@
     ],
     "require": {
         "ext-json": "*",
-        "php": "^7.2",
-        "illuminate/support": "~5.8.0||^6.0||^7.0",
-        "okipa/laravel-html-helper": "^1.3"
+        "php": "^7.4",
+        "illuminate/support": "^7.0||^8.0",
+        "okipa/laravel-html-helper": "^1.4"
     },
     "require-dev": {
-        "nunomaduro/larastan": "^0.4.3||^0.6",
-        "orchestra/testbench": "~3.8.0||^4.0||^5.0",
+        "nunomaduro/larastan": "^0.6",
+        "orchestra/testbench": "^5.0||^6.0",
         "phpmd/phpmd": "^2.8",
         "squizlabs/php_codesniffer": "^3.5"
     },
