@@ -4,33 +4,22 @@ namespace Okipa\LaravelBootstrapComponents\Components\Media\Abstracts;
 
 abstract class ImageAbstract extends MediaAbstract
 {
-    /** @property string $alt */
-    protected $alt;
+    protected ?string $alt = null;
 
-    /** @property int $width */
-    protected $width;
+    protected ?int $width = null;
 
-    /** @property int $height */
-    protected $height;
+    protected ?int $height = null;
 
-    /** @property string $linkId */
-    protected $linkId;
+    protected ?string $linkId = null;
 
-    /** @property array $linkClasses */
-    protected $linkClasses;
+    protected array $linkClasses = [];
 
-    /** @property string $linkUrl */
-    protected $linkUrl;
+    protected ?string $linkUrl = null;
 
-    /** @property string $linkTitle */
-    protected $linkTitle;
+    protected ?string $linkTitle = null;
 
-    /** @property array $linkHtmlAttributes */
-    protected $linkHtmlAttributes;
+    protected array $linkHtmlAttributes = [];
 
-    /**
-     * ImageAbstract constructor.
-     */
     public function __construct()
     {
         parent::__construct();
@@ -38,13 +27,6 @@ abstract class ImageAbstract extends MediaAbstract
         $this->linkHtmlAttributes = $this->setLinkHtmlAttributes();
     }
 
-    /**
-     * Define the image component alt HTML tag.
-     *
-     * @param string $alt
-     *
-     * @return $this
-     */
     public function alt(string $alt): self
     {
         $this->alt = $alt;
@@ -52,13 +34,6 @@ abstract class ImageAbstract extends MediaAbstract
         return $this;
     }
 
-    /**
-     * Define the component image HTML tag width.
-     *
-     * @param int $width
-     *
-     * @return $this
-     */
     public function width(int $width): self
     {
         $this->width = $width;
@@ -66,13 +41,6 @@ abstract class ImageAbstract extends MediaAbstract
         return $this;
     }
 
-    /**
-     * Define the component image HTML tag height.
-     *
-     * @param int $height
-     *
-     * @return $this
-     */
     public function height(int $height): self
     {
         $this->height = $height;
@@ -80,13 +48,6 @@ abstract class ImageAbstract extends MediaAbstract
         return $this;
     }
 
-    /**
-     * Set the image component link URL.
-     *
-     * @param string $linkUrl
-     *
-     * @return $this
-     */
     public function linkUrl(string $linkUrl): self
     {
         $this->linkUrl = $linkUrl;
@@ -94,13 +55,6 @@ abstract class ImageAbstract extends MediaAbstract
         return $this;
     }
 
-    /**
-     * Set the image component link title.
-     *
-     * @param string $linkTitle
-     *
-     * @return $this
-     */
     public function linkTitle(string $linkTitle): self
     {
         $this->linkTitle = $linkTitle;
@@ -108,13 +62,6 @@ abstract class ImageAbstract extends MediaAbstract
         return $this;
     }
 
-    /**
-     * Set the image component link id.
-     *
-     * @param string $linkId
-     *
-     * @return $this
-     */
     public function linkId(string $linkId): self
     {
         $this->linkId = $linkId;
@@ -122,13 +69,6 @@ abstract class ImageAbstract extends MediaAbstract
         return $this;
     }
 
-    /**
-     * Set the image component link classes.
-     *
-     * @param array $linkClasses
-     *
-     * @return $this
-     */
     public function linkClasses(array $linkClasses): self
     {
         $this->linkClasses = $linkClasses;
@@ -136,13 +76,6 @@ abstract class ImageAbstract extends MediaAbstract
         return $this;
     }
 
-    /**
-     * Set the image component link HTML attributes.
-     *
-     * @param array $linkHtmlAttributes
-     *
-     * @return $this
-     */
     public function linkHtmlAttributes(array $linkHtmlAttributes): self
     {
         $this->linkHtmlAttributes = $linkHtmlAttributes;
@@ -150,7 +83,6 @@ abstract class ImageAbstract extends MediaAbstract
         return $this;
     }
 
-    /** @inheritDoc */
     protected function getValues(): array
     {
         $alt = $this->getAlt();
@@ -168,97 +100,47 @@ abstract class ImageAbstract extends MediaAbstract
         );
     }
 
-    /**
-     * Get the image component alt attribute.
-     *
-     * @return string|null
-     */
     protected function getAlt(): ?string
     {
         return $this->alt ?: $this->getLabel() ?: $this->linkTitle;
     }
 
-    /**
-     * Get the image component width attribute.
-     *
-     * @return string|null
-     */
-    protected function getWidth(): ?string
+    protected function getWidth(): ?int
     {
         return $this->width;
     }
 
-    /**
-     * Get the image component height attribute.
-     *
-     * @return string|null
-     */
-    protected function getHeight(): ?string
+    protected function getHeight(): ?int
     {
         return $this->height;
     }
 
-    /**
-     * Set the default component id.
-     *
-     * @return string|null
-     */
     protected function getLinkId(): ?string
     {
         return $this->linkId;
     }
 
-    /**
-     * Set the default container class.
-     *
-     * @return array
-     */
     protected function getLinkClasses(): array
     {
         return $this->linkClasses;
     }
 
-    /**
-     * Set the image component link classes.
-     *
-     * @return array
-     */
     abstract protected function setLinkClasses(): array;
 
-    /**
-     * Get the image component link url.
-     *
-     * @return string|null
-     */
     protected function getLinkUrl(): ?string
     {
         return $this->linkUrl;
     }
 
-    /**
-     * Get the image component link url.
-     *
-     * @return string|null
-     */
     protected function getLinkTitle(): ?string
     {
         return $this->linkTitle ?: $this->getLabel() ?: $this->alt;
     }
 
-    /**
-     * Set the default component HTML attributes.
-     *
-     * @return array
-     */
     protected function getLinkHtmlAttributes(): array
     {
         return $this->linkHtmlAttributes;
     }
 
-    /**
-     * Set the image component link HTML attributes.
-     *
-     * @return array
-     */
     abstract protected function setLinkHtmlAttributes(): array;
 }
