@@ -4,25 +4,14 @@ namespace Okipa\LaravelBootstrapComponents\Components\Media\Abstracts;
 
 abstract class VideoAbstract extends MediaAbstract
 {
-    /** @property string $poster */
-    protected $poster;
+    protected ?string $poster;
 
-    /**
-     * VideoAbstract constructor.
-     */
     public function __construct()
     {
         parent::__construct();
         $this->poster = $this->setPoster();
     }
 
-    /**
-     * Set the video component poster.
-     *
-     * @param string $poster
-     *
-     * @return $this
-     */
     public function poster(string $poster): self
     {
         $this->poster = $poster;
@@ -30,7 +19,6 @@ abstract class VideoAbstract extends MediaAbstract
         return $this;
     }
 
-    /** @inheritDoc */
     protected function getValues(): array
     {
         $poster = $this->getPoster();
@@ -38,20 +26,10 @@ abstract class VideoAbstract extends MediaAbstract
         return array_merge(parent::getValues(), compact('poster'));
     }
 
-    /**
-     * Set the default video poster.
-     *
-     * @return string|null
-     */
     protected function getPoster(): ?string
     {
         return $this->poster;
     }
 
-    /**
-     * Set the video component poster path.
-     *
-     * @return string|null
-     */
     abstract protected function setPoster(): ?string;
 }
