@@ -6,12 +6,7 @@ use Exception;
 
 trait FormValidityChecks
 {
-    /**
-     * Check the component values validity
-     *
-     * @return void
-     * @throws \Exception
-     */
+    /** @throws \Exception */
     protected function checkValuesValidity(): void
     {
         if (! $this->getName()) {
