@@ -6,10 +6,7 @@ use Illuminate\Support\Facades\Facade;
 
 class Video extends Facade
 {
-    /**
-     * {@inheritdoc}
-     */
-    protected static function getFacadeAccessor()
+    protected static function getFacadeAccessor(): string
     {
         return 'Video';
     }
